work on async and reactivity

+++ tests/test-pg-database.js
@@ -5,11 +5,11 @@ const { PGDatabase }	= require('../index');
 const { PGTable }		= require('../lib/pg-table');
 const { PGView }		= require('../lib/pg-view');
 
-const DB_USER = 'postgres';
-const DB_HOST = 'localhost';
+const DB_HOST = '127.0.0.1';
 const DB_DATABASE = 'pureworkx';
+const DB_USER = 'pureworkx';
 const DB_PASSWORD = 'testtest';
-const DB_PORT = 5432;
+const DB_PORT = /*version 10.0*/ 5433; // or port /*version 9.5*/ 5432;
 const DB_MAX_CONNECTION = 100;
 
 describe('PgDatabase', function() {
@@ -228,14 +228,13 @@ describe('PgDatabase', function() {
 						expect(result.rows[0].email).to.equal('jrysdaleke@vistaprint.com');
 
 						db.end();
-
 						done();
 					});
 				});
 			});
 		});
 
-		it('should reactively get records from core_users', function(done) {
+		it('should reactively get records from people', function(done) {
 			let db = new PGDatabase({
 				enablePooling: true,
 				connect: {
@@ -265,6 +264,7 @@ describe('PgDatabase', function() {
 							expect(peopleCounter).to.equal(2);
 
 							reactiveQuery.destroy(function(error){
+								db.end();
 								done();
 							});
 						}
@@ -299,6 +299,7 @@ describe('PgDatabase', function() {
 						]
 					}, (error, result) => {
 						if (error) return done(error);
+						db.end();
 						done();
 					});
 				});
@@ -329,6 +330,7 @@ describe('PgDatabase', function() {
 						email: 'tester0.test@test.com.de'
 					}, (error, result) => {
 						if (error) return done(error);
+						db.end();
 						done();
 					});
 				});
@@ -368,6 +370,7 @@ describe('PgDatabase', function() {
 						if (error) return done(error);
 
 						expect(result.rowCount).to.equal(2);
+						db.end();
 						done();
 					});
 				});
@@ -399,54 +402,68 @@ describe('PgDatabase', function() {
 						if (error) return done(error);
 
 						expect(result.rowCount).to.equal(3);
+						db.end();
 						done();
 					});
 				});
 			});
 		});
 
-		it('should reactively update one record from core_users and get a change event', function(done) {
-			// remove the query preparators first
-			db._queryPreparators = [];
+		it('should reactively update one record from people and get a change event', function(done) {
+			let db = new PGDatabase({
+				enablePooling: true,
+				connect: {
+					user: DB_USER,
+					host: DB_HOST,
+					database: DB_DATABASE,
+					password: DB_PASSWORD,
+					port: DB_PORT,
+					max: DB_MAX_CONNECTION
+				}
+			});
 
-			var Users = db.Table('core_users');
-			// update the emailadresse before running the reactive query
-			// to have a change on the data with the next update
-			Users.update({_id: '5fde3570-8fb2-11e7-a232-bc307d530814'}, {
-				emailaddress: 'XXX0-newmailadr@gmail.com'
-			}, function(error, result){
-				expect(error).to.equal(null);
+			db.connect((error) => {
+				var People = db.Table('public.people', (error) => {
+					if (error) return done(error);
 
-				var reactiveQuery = Users.find({
-					created_by: 'peter'
-				});
-
-				reactiveQuery.on('changed', (id, row) => {
-					expect(id).to.equal('5fde3570-8fb2-11e7-a232-bc307d530814');
-					expect(row.emailaddress).to.equal('newmailadr0@gmail.com');
-
-					reactiveQuery.destroy( () => {
-						done();
+					var reactiveQuery = People.find({
+						first_name: 'Tester'
 					});
+
+					var changeCounter = 0;
+					reactiveQuery.on('changed', (id, row) => {
+						console.log('changed', id);
+						changeCounter++;
+
+						expect(row.email).to.equal('newmailadr0@gmail.com');
+
+						if (changeCounter == 3) {
+							console.log('3 changes occured');
+							reactiveQuery.destroy( () => {
+								db.end();
+								done();
+							});
+						}
+					});
+
+					var firstTime = true;
+					reactiveQuery.on('state', (currentState) => {
+						if (currentState == 'ready'){
+							if (!firstTime) return;
+							firstTime = false;
+
+							People.update({ first_name: 'Tester'}, {
+								email: 'newmailadr0@gmail.com'
+							}, function(error, result){
+								console.log('UPDATED.');
+								expect(error).to.equal(null);
+								expect(result.rowCount).to.equal(3);
+							});
+						}
+					});
+
+					reactiveQuery.run();
 				});
-
-				var firstTime = true;
-				reactiveQuery.on('state', (currentState) => {
-					if (currentState == 'ready'){
-						if (!firstTime) return;
-						firstTime = false;
-
-						Users.update({_id: '5fde3570-8fb2-11e7-a232-bc307d530814'}, {
-							emailaddress: 'newmailadr0@gmail.com'
-						}, function(error, results){
-							expect(error).to.equal(null);
-
-						});
-
-					}
-				});
-
-				reactiveQuery.run();
 			});
 		});
 	});
